Compatibility with 'Sensible Entrance Points' from Tweaks Anthology
On Tutu, the new world map was overwriting the new entrance names added by Tweaks Anthology. If this component from Anthology is detected, the links will be updated to the new Anthology entrance names, removing the incompatibility and ensuring both work as intended.

Since BGT already had defined entrance names, no actual compatibility code was required and it has been removed. The BGT framework was not used as a template for the Tutu compatibility code, as the arrays were overwriting themselves with non-unique keys.